Satellite Menu
'Path' has a very attractive menu sitting on the left bottom corner of
the screen. Satellite Menu is the open version of this menu.

For the ones who didnt see path, the menu consists of a main button on
the left bottom of the screen. When the user clicks on this button, a
number of buttons are popped out of this button. I have resembled this
structure to a planet and satellites and thats why I gave the name of
this project.